// server/vance-addon-brain-gtd/client/src/generated/gtd/GtdView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.gtd.GtdView
*/

import { GtdBucketView } from './GtdBucketView';
import { GtdStatsView } from './GtdStatsView';
import { GtdProjectView } from './GtdProjectView';

export interface GtdView {
  folder: string;
  title?: string;
  description?: string;
  contexts: string[];
  buckets: GtdBucketView[];
  projects: GtdProjectView[];
  stats: GtdStatsView;
}