// server/vance-addon-brain-gtd/client/src/generated/gtd/GtdActionContentView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.gtd.GtdActionContentView
*/

export interface GtdActionContentView {
  id: string;
  path: string;
  title: string;
  when: string;
  deadline?: string;
  contexts: string[];
  done: boolean;
  project?: string;
  body: string;
}