// server/vance-addon-brain-gtd/client/src/generated/gtd/GtdActionRequest.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.gtd.GtdActionRequest
*/

export interface GtdActionRequest {
  title?: string;
  when?: string;
  deadline?: string;
  contexts?: string[];
  project?: string;
  done?: boolean;
  body?: string;
}